feat(runtime): Refactor skill runtime structure and introduce LocalRuntime implementation
